refactor(server): subpackage for step
Easier to have one file per step
