+++ server/weaverbird/types.py
@@ -1,0 +1,5 @@
+from typing import Callable
+
+from pandas import DataFrame
+
+DomainRetriever = Callable[[str], DataFrame]
